Pruebas de integracion

+++ src/main/java/market/rest/CategoryRest.java
@@ -1,6 +1,10 @@
 package market.rest;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,6 +14,7 @@ import market.model.Category;
 import market.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,11 +27,16 @@ public class CategoryRest {
     private CategoryService categoryService;
 
     @PostMapping
-    public ResponseEntity<Category>
+    public ResponseEntity<?>
             save(@RequestBody Category c) {
-        Category result = categoryService.save(c);
-        return ResponseEntity.ok(result);
-
+        try {
+            Category result = categoryService.save(c);
+            return ResponseEntity.
+                    created(new URI("/category/"+result.getId()))
+                    .body(result);
+        } catch (URISyntaxException | RuntimeException ex) {
+            return ResponseEntity.internalServerError().body("Error");
+        }
     }
             
     @GetMapping
@@ -46,9 +56,15 @@ public class CategoryRest {
     }
     
     @DeleteMapping("{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable long id){
-        categoryService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<?> deleteById(@PathVariable long id){
+        Category result = categoryService.getById(id);
+        if(result!=null){
+            categoryService.deleteById(id);
+            return ResponseEntity.ok().build();    
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Categoria no encontrada");
+        }
     }
 
 }
